refactor(#233): hoist dispatcher-workspace helper to service root
Fix the layering inversion flagged in review: the shared `worktree/`
helper imported `ensureDispatcherWorkspace` from `dispatcher-service/`,
i.e. a cross-cutting helper reaching up into a specific service.

`dispatcher-workspace.ts` (the issue #182 dispatcher-cwd policy) is consumed
by server-startup preflight, the dispatcher service, `dreamux doctor`, and the
worktree layer — no single service owns it. Move it to the `service/` root
beside the other cross-cutting helpers (`worktree/`, `channel-binding/`,
`legacy-state.ts`); update all four importers and the docs.

No behavior change. tsc 0, lint 0.
